moved incrementPlayerQuest and cleaned code
