+++ controllers/questsController.js
@@ -138,4 +138,45 @@ function updatePlayerQuest(io, socket, { user, questTitle, newObjectiveRef }) {
     }
 }
 
-module.exports = { assignAndUpdatePlayerQuest, updatePlayerQuest }
+function incrementPlayerQuest(io, socket, { questToUpdate: { title, objectiveReference, completed }, user: { characterName, quests, tokens } }) {
+    try {
+        !completed && db.Quest.findOne({ title }).then(data => {
+            data = data.toJSON();
+            let oldIndex = data.objectives.findIndex(objective => objective.reference === objectiveReference);
+            let oldObjective = data.objectives[oldIndex];
+            let newIndex = oldIndex + 1;
+            let newObjective = data.objectives[newIndex];
+            let completed = data.objectives.length - 1 === newIndex
+            let updatedPlayerQuest = {
+                title,
+                objectiveReference: newObjective.reference,
+                completed
+            }
+            quests = quests.map(quest => {
+                if (quest.title === title) {
+                    return updatedPlayerQuest
+                } else {
+                    return quest
+                }
+            })
+            db.Player.findOneAndUpdate({ characterName }, { quests }, { new: true })
+                .then(({ quests }) => {
+                    io.to(socket.id).emit('questsUpdate', { quests })
+                    if (completed) {
+                        io.to(socket.id).emit('completedQuest')
+                    } else {
+                        io.to(socket.id).emit('updatedQuest')
+                    }
+                })
+                .catch(e => {
+                    console.log("ERROR IN DB CALL", e)
+                    io.to(socket.id).emit('failure', 'Something went wrong')
+                })
+        })
+    } catch (e) {
+        console.log("ERROR IN incremenetPlayerQuest", e)
+        io.to(socket.id).emit('failure', 'Something went wrong')
+    }
+}
+
+module.exports = { assignAndUpdatePlayerQuest, updatePlayerQuest, incrementPlayerQuest }
